+++ plugins/tasks/ansible/src/main/resources/com/walmartlabs/concord/plugins/ansible/callback/concord_out_vars.py
@@ -1,5 +1,5 @@
-import ujson as json
 import os
+import ujson as json
 from ansible.plugins.callback import CallbackBase
 from concord_ansible_stats import ConcordAnsibleStats