+++ plugins/tasks/ansible/src/main/resources/com/walmartlabs/concord/plugins/ansible/callback/concord_trace.py
@@ -1,6 +1,6 @@
-import ujson as json
-import os
 import errno
+import os
+import ujson as json
 from ansible.plugins.callback import CallbackBase
 from concord_ansible_stats import ConcordAnsibleStats