+++ plugins/tasks/ansible/src/main/resources/com/walmartlabs/concord/plugins/ansible/lookup/concord_secret.py
@@ -2,11 +2,7 @@ from __future__ import (absolute_import, division, print_function)
 
 __metaclass__ = type
 
-from ansible.errors import AnsibleError
-from ansible.plugins.lookup import LookupBase
-
 import os
-import requests
 import sys
 
 sys.path.insert(1, os.path.join(os.path.dirname(__file__), os.pardir, "_lookups"))