+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - concord-server, concord-agents: external `imports` are now
 processed without copying into the process state;
+- ansible: send process events asynchronously;
 - concord-client, concord-runner: use a custom `User-Agent` header;
 - concord-server: fixed a NPE when handling optional `file` form
 fields.